Started lab 3. Continue from RemoveTags.cc implementation. The tests are probably wrong but should be some quick fixes with spacing and so on bro

+++ lab2/edit_distance.cc
@@ -9,14 +9,14 @@ int d[Dictionary::maxlen + 1][Dictionary::maxlen + 1];
 
 int edit_distance(const string &word, const string &spelled_word)
 {
-    for (size_t i = 0; i <= word.size(); ++i) 
+    for (int i = 0; i <= word.size(); ++i) 
         d[i][0] = i;
-    for (size_t j = 0; j <= spelled_word.size(); ++j) 
+    for (int j = 0; j <= spelled_word.size(); ++j) 
         d[0][j] = j;
 
-    for (size_t i = 1; i <= word.size(); ++i)
+    for (int i = 1; i <= word.size(); ++i)
     {
-        for (size_t j = 1; j <= spelled_word.size(); ++j)
+        for (int j = 1; j <= spelled_word.size(); ++j)
         {
             int cost = (spelled_word[j - 1] == word[i - 1]) ? 0 : 1;
             d[i][j] = std::min(std::min(d[i - 1][j] + 1, d[i][j - 1] + 1), d[i - 1][j - 1] + cost);
